# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cocoon backend API
  summary: LangGraph application backend for the Cocoon voice assistant (Team Butterfly).
  version: 1.0.0
paths:
...
  /v1/me:
    get:
      tags:
      - identity
      summary: Me
      operationId: me_v1_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: 'unauthorized: missing, malformed, unknown, expired or revoked bearer token, or more than
            one Authorization header (the cause is deliberately not distinguished)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'forbidden: authenticated, but this role may not use the route'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Unknown session or resource; also returned for a session an operator does not own
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (retryable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'auth_unavailable: the token store could not be read (retryable; never treated as success)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /v1/sessions:
    post:
      tags:
      - sessions
      summary: Create Session
      operationId: create_session_v1_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          description: 'unauthorized: missing, malformed, unknown, expired or revoked bearer token, or more than
            one Authorization header (the cause is deliberately not distinguished)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'forbidden: authenticated, but this role may not use the route'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Unknown session or resource; also returned for a session an operator does not own
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Malformed request (validation_error), or a NEW session whose machine_id (unknown_machine,
            checked first) or operator_id (unknown_operator) is not in the verified catalog, or whose site/shift
            has no trusted binding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (retryable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'catalog_unavailable: no verified catalog loaded; new sessions refused; or auth_unavailable:
            the token store could not be read (retryable; never treated as success)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '200':
          description: Existing session for this client_session_key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '409':
          description: Key already bound to a different room/participant/operator/machine/site/shift
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - HTTPBearer: []
  /v1/sessions/{session_id}/turns:
    post:
      tags:
      - turns
      summary: Submit Turn
      operationId: submit_turn_v1_sessions__session_id__turns_post
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TurnRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnResult'
        '401':
          description: 'unauthorized: missing, malformed, unknown, expired or revoked bearer token, or more than
            one Authorization header (the cause is deliberately not distinguished)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'forbidden: authenticated, but this role may not use the route'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Unknown session or resource; also returned for a session an operator does not own
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (retryable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'LLM unavailable or turn timed out (retryable, same turn_id); or auth_unavailable: the token
            store could not be read (retryable; never treated as success)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '202':
          description: Same turn_id is still processing. Wait retry_after_ms (also the Retry-After header, seconds)
            then GET poll_url (also the Location header). Do not resubmit with a new turn_id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnResult'
        '409':
          description: turn_id reused with a different text or source
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/sessions/{session_id}/turns/{turn_id}:
    get:
      tags:
      - turns
      summary: Get Turn
      operationId: get_turn_v1_sessions__session_id__turns__turn_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: turn_id
        in: path
        required: true
        schema:
          type: string
          title: Turn Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurnResult'
        '401':
          description: 'unauthorized: missing, malformed, unknown, expired or revoked bearer token, or more than
            one Authorization header (the cause is deliberately not distinguished)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'forbidden: authenticated, but this role may not use the route'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Unknown session or resource; also returned for a session an operator does not own
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (retryable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'auth_unavailable: the token store could not be read (retryable; never treated as success)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/sessions/{session_id}/commands:
    post:
      tags:
      - commands
      summary: Submit Command
      operationId: submit_command_v1_sessions__session_id__commands_post
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionCommand'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCommandResult'
        '401':
          description: 'unauthorized: missing, malformed, unknown, expired or revoked bearer token, or more than
            one Authorization header (the cause is deliberately not distinguished)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'forbidden: authenticated, but this role may not use the route'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Unknown session or resource; also returned for a session an operator does not own
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (retryable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'auth_unavailable: the token store could not be read (retryable; never treated as success)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: idempotency_conflict (command_id reused with a different payload), version_conflict (expected_version
            is stale; details give the current version) or invalid_transition (the task is not in a state this command
            accepts)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/sessions/{session_id}/commands/{command_id}:
    get:
      tags:
      - commands
      summary: Get Command
      operationId: get_command_v1_sessions__session_id__commands__command_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: command_id
        in: path
        required: true
        schema:
          type: string
          title: Command Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCommandResult'
        '401':
          description: 'unauthorized: missing, malformed, unknown, expired or revoked bearer token, or more than
            one Authorization header (the cause is deliberately not distinguished)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 'forbidden: authenticated, but this role may not use the route'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Unknown session or resource; also returned for a session an operator does not own
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Malformed request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error (retryable)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: 'auth_unavailable: the token store could not be read (retryable; never treated as success)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      additionalProperties: false
      type: object
      required:
      - error
      title: ErrorResponse
...
    TurnRequest:
      properties:
        turn_id:
          type: string
          pattern: ^[A-Za-z0-9][A-Za-z0-9._:\-]{0,127}$
          title: Turn Id
          examples:
          - lk-item_4f2c9a
        text:
          type: string
          maxLength: 2000
          minLength: 1
          title: Text
        source:
          type: string
          enum:
          - voice
          - text
          title: Source
      additionalProperties: false
      type: object
      required:
      - turn_id
      - text
      - source
      title: TurnRequest
...
  securitySchemes:
    HTTPBearer:
      type: http
      description: 'Exactly ONE of: the trusted service credential (COCOON_SERVICE_TOKEN; server-side callers only,
        never embedded in Android or React), or an actor token (`cct_...`) issued locally by scripts/actor_tokens.py
        for an operator or supervisor principal. Access per route is listed in API_CONTRACT.md.'
      scheme: bearer